Fix occasional start bug.
The fix is similar to the one in commit e21cf13d6bdb92428975b2cc702a2a6a38bd3b34. The problem is that the variable `time` is 0 only for a very small time, and LiveSplit doesn't update fast enough to read it on slower computers. Because of this, `if (wentToNextLevel && time == 0)` has been changed too.

In ShouldStart, `if (currentLevel >= Level.DemoVenice || currentLevel == Level.LarasHome)` has been changed to `if (currentLevel >= Level.DemoVenice`, because the game initializes `currentLevel` to `Level.LarasHome` and that prevents the new game select check from running. Previously, this wasn't an issue, because `currentLevel` was `Level.TheGreatWall` when the new game check code ran (and there were no select check).